+++ sixel.h
@@ -20,6 +20,7 @@ void sixel_scroll_up(struct terminal *term, int rows);
 void sixel_scroll_down(struct terminal *term, int rows);
 
 void sixel_cell_size_changed(struct terminal *term);
+void sixel_sync_cache(const struct terminal *term, struct sixel *sixel);
 
 void sixel_reflow_grid(struct terminal *term, struct grid *grid);